# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CivicAI – Smart Civic Issue Reporting Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for the authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required: [ok]
...
  /users/me:
    get:
      operationId: getCurrentUser
      tags: [users]
      summary: Current authenticated user (JWT)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users:
    post:
      operationId: createUser
      tags: [users]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/issues:
    get:
      operationId: getUserIssues
      tags: [users]
      summary: List issues reported by a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of user issues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Issue"
  /users/login:
    post:
      operationId: loginUser
      tags: [users]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials

  # ─── LEADERBOARD ────────────────────────────────────────────────────────────
  /users/leaderboard:
    get:
      operationId: getLeaderboard
      tags: [users]
      summary: Get top users by points
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Leaderboard
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"

  # ─── REWARDS ────────────────────────────────────────────────────────────────
  /rewards/{userId}:
    get:
      operationId: getUserRewards
      tags: [rewards]
      summary: Get user rewards and wallet
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User rewards
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRewards"
  /rewards/withdraw:
    post:
      operationId: withdrawRewards
      tags: [rewards]
      summary: Request UPI withdrawal (mock)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawBody"
      responses:
        "200":
          description: Withdrawal initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawResponse"

  # ─── AUTHORITIES ─────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT from /users/login, /authorities/login, or /auth/verify-otp
  schemas:
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        category:
          type: string
        title:
          type: string
        body:
          type: string
        read:
          type: boolean
        issueId:
          type: integer
          nullable: true
        metadata:
          type: object
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, category, title, body, read, createdAt]
...
    Issue:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        authorityId:
          type: integer
          nullable: true
        issueType:
          type: string
          enum: [garbage, pothole, water_leakage, street_light, plastic, other]
        description:
          type: string
        imageUrl:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        reporterLatitude:
          type: number
          nullable: true
        reporterLongitude:
          type: number
          nullable: true
        address:
          type: string
        status:
          type: string
          enum: [pending, in_progress, resolved]
        verificationStatus:
          type: string
          enum: [pending, flagged, rejected, approved]
        isDuplicate:
          type: boolean
        isValid:
          type: boolean
        validationNotes:
          type: string
          nullable: true
        department:
          type: string
        confidenceScore:
          type: number
        resolvedImageUrl:
          type: string
          nullable: true
        imageHash:
          type: string
        imageMd5:
          type: string
          nullable: true
        imagePhash:
          type: string
          nullable: true
        exifTakenAt:
          type: string
          format: date-time
          nullable: true
        exifLatitude:
          type: number
          nullable: true
        exifLongitude:
          type: number
          nullable: true
        reporterName:
          type: string
          nullable: true
        pointsAwarded:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - issueType
        - latitude
        - longitude
        - status
        - verificationStatus
        - isDuplicate
        - isValid
        - department
        - imageHash
        - pointsAwarded
        - createdAt
        - updatedAt
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        emailVerified:
          type: boolean
        phone:
          type: string
          nullable: true
        role:
          type: string
          enum: [user, authority, admin]
        points:
          type: integer
        badge:
          type: string
          nullable: true
        totalReports:
          type: integer
        resolvedReports:
          type: integer
        validReports:
          type: integer
        rejectedReports:
          type: integer
        trustScore:
          type: number
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, points, totalReports, resolvedReports, validReports, rejectedReports, trustScore, createdAt]
    CreateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: string
      required: [name, email, password]
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
...
    LeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
        userId:
          type: integer
        name:
          type: string
        points:
          type: integer
        badge:
          type: string
          nullable: true
        totalReports:
          type: integer
        resolvedReports:
          type: integer
      required: [rank, userId, name, points, totalReports, resolvedReports]
    UserRewards:
      type: object
      properties:
        userId:
          type: integer
        totalPoints:
          type: integer
        cashValue:
          type: number
        pendingWithdrawal:
          type: number
        totalWithdrawn:
          type: number
        badge:
          type: string
          nullable: true
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/RewardTransaction"
      required: [userId, totalPoints, cashValue, pendingWithdrawal, totalWithdrawn, transactions]
    RewardTransaction:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [earned, withdrawn]
        points:
          type: integer
        description:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, type, points, description, createdAt]
    WithdrawBody:
      type: object
      properties:
        userId:
          type: integer
        points:
          type: integer
        upiId:
          type: string
      required: [userId, points, upiId]
    WithdrawResponse:
      type: object
      properties:
        success:
          type: boolean
        transactionId:
          type: string
        amount:
          type: number
        message:
          type: string
      required: [success, transactionId, amount, message]
...